Better handling of defaults for other types.
This is because Postgres doesn't seem to easily allow passing an argument for `default` in `add column` , and I can't find a quote value function. Also, that means the eventual `autoresponse_file` default won't have to be stored for each server in the table, although it might not even be stored like that Idk.